create separate interpreters and helpers to split functionalities of keccak
